Add function to check distance, and add form for product details
Sam, James and Howard worked on making an acceptance system that only allows the user to scan a product if they are within the range of a product. If not in range the submit button will update and a google maps to the nearest bin opens.

Oscar worked on creating a a form so that if the product the user scans is not in the database the user can input the weight and material of the product.

It was such fun.

Co-Authored-By: Howard Cui <78500597+HowardCui-Super@users.noreply.github.com>
Co-Authored-By: James Ford <57077523+Jamesemf@users.noreply.github.com>
Co-Authored-By: Oscar Norris <57996381+oscarhfnorris@users.noreply.github.com>