Transactions read to record
As a team, Sam, Howard and James; made the backend function which collects the barcode from a product and then determines whether the user should input data as it doesn't exist, or can record the event as it is already in the database. There is a form made by Oscar which allows the user to input the data on the product.

Oscar focused on making the form and handling the form request. He is wanting to rework the form so that users are quick entering information and are not annoyed by doing something.

Co-Authored-By: Howard Cui <78500597+HowardCui-Super@users.noreply.github.com>
Co-Authored-By: James Ford <57077523+Jamesemf@users.noreply.github.com>
Co-Authored-By: Oscar Norris <57996381+oscarhfnorris@users.noreply.github.com>