Goals, recent transactions
Together we added better descriptions to the Goals, so that the user can make custom goals from fixed templates. Coupled with limiting the amount of transactions that can been seen on the home page.

We also fixed the confirm password part of registration so that the passwords must match.

Improved security

Co-Authored-By: James Ford <57077523+Jamesemf@users.noreply.github.com>